requirements.txt
# Bybit Trading Bot Requirements
# Python 3.9+

# Telegram Bot
python-telegram-bot==20.7
python-telegram-bot[job-queue]==20.7

# Trading & Exchange APIs
ccxt==4.2.25
pybit==5.6.2

# Web Framework for Webhooks
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Database
aiosqlite==0.19.0

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0
aiohttp==3.9.1

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Utilities
pydantic==2.5.3
ujson==5.9.0

# Logging
loguru==0.7.2

# Async Support
asyncio==3.4.3

# UUID Generator
uuid==1.30

